pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rubot"
version = "0.1.0"
description = "CLI tool for downloading and processing Rathaus-Umschau PDFs"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "rmoriz", email = "rmoriz@users.noreply.github.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
requires-python = ">=3.13"
dependencies = [
    "click>=8.1.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-mock>=3.11.0",
    "pytest-click>=1.1.0",
    "pytest-cov>=4.0.0",
    "coverage>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[project.scripts]
rubot = "rubot.cli:main"

[project.urls]
Homepage = "https://github.com/rmoriz/rubot"
Repository = "https://github.com/rmoriz/rubot"
Issues = "https://github.com/rmoriz/rubot/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["rubot*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.black]
line-length = 88
target-version = ['py313']

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true